Generate TLS-PSK key on update
- mosq_readconfig restarts Mosquitto if not running
- updateconfig generates a 240bit TLS-PSK key on update

Mosquitto uses
use_identity_as_user=false
Therefore not required to give away user/pass to the bridge

+++ bin/sudo/mosq_readconfig.sh
@@ -1,2 +1,10 @@
 #!/bin/bash
-pkill -HUP mosquitto
+
+pgrep mosquitto
+if [ $? -ne 0 ]; then
+	echo "Restarting Mosquitto"
+	systemctl restart mosquitto
+else
+	echo "Re-reading Mosquitto config"
+	pkill -HUP mosquitto
+fi
